# specs/001-repair-operations-platform/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Repair Operations Platform API
  version: 0.1.0
  description: Contract for the MVP repair workflow, RBAC, evidence, quotations, payments, notifications, and reports.
servers:
  - url: https://app.example.local/api
security:
  - sessionAuth: []
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    JobStatus:
      type: string
      enum:
        - NEW JOB
        - WAITING FADHIL REVIEW
        - WAITING CUSTOMER CONFIRMATION
        - IN PROGRESS
        - READY PICKUP
        - COMPLETE
        - CANCELLED
        - CLOSED_NO_REPAIR
...
    Job:
      type: object
      required: [id, jobNumber, status, customerId, deviceId]
      properties:
        id: { type: string }
        jobNumber: { type: string }
        status: { $ref: '#/components/schemas/JobStatus' }
        customerId: { type: string }
        deviceId: { type: string }
        issueDescription: { type: string }
        assignedTechnicianId: { type: string, nullable: true }
        nasFolderPath: { type: string }
    CreateJobRequest:
      type: object
      required: [customer, device, issueDescription]
      properties:
        customer:
          type: object
          required: [name, phone]
          properties:
            name: { type: string }
            phone: { type: string }
            email: { type: string, nullable: true }
            preferredChannel: { type: string, enum: [whatsapp, email, phone, none] }
            whatsappConsent: { type: boolean }
            emailConsent: { type: boolean }
        device:
          type: object
          required: [type, brand]
          properties:
            type: { type: string }
            brand: { type: string }
            model: { type: string, nullable: true }
            serialNumber: { type: string, nullable: true }
            accessoriesReceived: { type: string, nullable: true }
            conditionNotes: { type: string, nullable: true }
        issueDescription: { type: string }
        intakeNotes: { type: string, nullable: true }
        serviceNoteExtract:
          $ref: '#/components/schemas/ServiceNoteScanResult'
    ServiceNoteScanResult:
      type: object
      properties:
        serviceNumber: { type: string, nullable: true }
        receivedDate: { type: string, nullable: true }
        customer:
          type: object
          properties:
            name: { type: string, nullable: true }
            phone: { type: string, nullable: true }
            email: { type: string, nullable: true }
        device:
          type: object
          properties:
            type: { type: string, nullable: true }
            brand: { type: string, nullable: true }
            model: { type: string, nullable: true }
            serialNumber: { type: string, nullable: true }
            specifications: { type: string, nullable: true }
            password: { type: string, nullable: true }
        repair:
          type: object
          properties:
            reportedIssue: { type: string, nullable: true }
            initialDiagnosis: { type: string, nullable: true }
            repairActionTaken: { type: string, nullable: true }
            partsReplaced: { type: string, nullable: true }
        charges:
          type: object
          properties:
            serviceFee: { type: string, nullable: true }
            partsFee: { type: string, nullable: true }
            total: { type: string, nullable: true }
        warranty:
          type: object
          properties:
            days: { type: string, nullable: true }
        confidenceByField:
          type: object
          additionalProperties: { type: number }
        rawTextNotes: { type: string, nullable: true }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
paths:
  /service-notes/scan:
    post:
      summary: Extract service-note fields from an uploaded PDF or image
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Extracted service-note draft
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ServiceNoteScanResult' }
        '400':
          description: Missing or unsupported file
        '502':
          description: Extraction provider unavailable or invalid output
  /jobs:
    get:
      summary: List jobs by status, branch, customer, or assigned technician
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/JobStatus' }
        - in: query
          name: q
          schema: { type: string }
      responses:
        '200':
          description: Job list
    post:
      summary: Create a new repair job
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateJobRequest' }
      responses:
        '201':
          description: Job created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '403':
          description: Role not allowed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /jobs/{jobId}:
    get:
      summary: Get job detail with timeline and permitted evidence metadata
      parameters:
        - in: path
          name: jobId
          required: true
          schema: { type: string }
      responses:
        '200': { description: Job detail }
...
  /reports/jobs:
    get:
      summary: Job and payment report by date range
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
      responses:
        '200': { description: Report data }
...